[feat] 보스 몬스터 필수 기능 구현 완료

+++ Assets/_Scripts/Monster/Core/Base/NormalMonster.cs
@@ -113,7 +113,7 @@ public abstract class NormalMonster : BaseMonster, Game.Monster.IDamageable
         _sr.material = _originMat;
     }
 
-    public abstract void TakeDamage(int damage);
+    public abstract void TakeDamage(float damage);
 
     /// <summary>
     /// 해당 콜라이더를 충돌 무시로 지정하고 리스트에 넣음
